Turn tapas into a tsapp app.
This was straightfoward except for one major change:

The content which was in the tapas_extra bag is now in the
tapas_public bag. In order for the proxy to work, the requested
content cannot have a full URI with, yet without that, the
non-space bags cannot be reached, once deployed, because of
ControlView.

So for now, just not using extra bags.

ControlView is silly. Maybe it should allow anything with
<space>_ prefix?