fix #6 and #7
This allows tapas to access the socket server on port 8080 instead
of 8081, it also makes sure that the hostname is not used in the
javascript files, instead we get the proper host from /status.js.

However there remains one static reference to tiddlyspace.com in
tapas.html which will require some discussion with @pads.